Move clearSelection from util to device.
