+++ src/device/clearSelection.ts
@@ -1,7 +1,9 @@
 /** Clears the browser selection. */
-export const clearSelection = (): void => {
+const clearSelection = (): void => {
   const sel = window.getSelection()
   if (sel) {
     sel.removeAllRanges()
   }
 }
+
+export default clearSelection
