+++ src/util/index.ts
@@ -5,7 +5,6 @@ export { addEmojiSpace } from './addEmojiSpace'
 export { ancestors } from './ancestors'
 export { appendToPath } from './appendToPath'
 export { checkIfPathShareSubcontext } from './checkIfPathShareSubcontext'
-export { clearSelection } from './clearSelection'
 export { compareByRank } from './compareByRank'
 export { compareThought, compareThoughtDescending } from './compareThought'
 export { componentToThought } from './componentToThought'
