+++ src/util/initEvents.native.ts
@@ -2,11 +2,12 @@ import _ from 'lodash'
 import { Store } from 'redux'
 import { inputHandlers /* isGestureHint */ } from '../shortcuts'
 import * as db from '../data-providers/dexie'
-import { clearSelection, isRoot, pathToContext } from '.'
+import { isRoot, pathToContext } from '.'
 import { decodeThoughtsUrl, pathExists } from '../selectors'
 import { /* alert */ error, setCursor, toggleTopControlsAndBreadcrumbs } from '../action-creators'
 import scrollCursorIntoView from '../device/scrollCursorIntoView'
 import { equalPath } from './equalPath'
+import clearSelection from '../device/clearSelection'
 import { Path, State } from '../@types'
 // import lifecycle from 'page-lifecycle'
