+++ src/util/initEvents.ts
@@ -2,10 +2,11 @@ import _ from 'lodash'
 import { Store } from 'redux'
 import { inputHandlers, isGestureHint } from '../shortcuts'
 import * as db from '../data-providers/dexie'
-import { clearSelection, isRoot, pathToContext } from '../util'
+import { isRoot, pathToContext } from '../util'
 import { decodeThoughtsUrl, pathExists } from '../selectors'
 import { alert, error, setCursor, toggleTopControlsAndBreadcrumbs } from '../action-creators'
 import scrollCursorIntoView from '../device/scrollCursorIntoView'
+import clearSelection from '../device/clearSelection'
 import { Path, State } from '../@types'
 import { equalPath } from './equalPath'
 import lifecycle from 'page-lifecycle'
